add websocket, rabbitmq, add refreshtoken, restructure src code

+++ backend/src/main/java/com/example/sep490/repositories/UserRepository.java
@@ -8,7 +8,7 @@ import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 
 import java.util.Optional;
 
-public interface UserRepository extends JpaRepository<User, Integer> , JpaSpecificationExecutor<User> {
+public interface UserRepository extends JpaRepository<User, Long> , JpaSpecificationExecutor<User> {
     Page<User> findByNameContainingIgnoreCaseAndIsDeleteFalse(String name, Pageable pageable);
     Page<User> findByIsDeleteFalse(Pageable pageable);
     Optional<User> findByName(String username);
